BiddingPlatform/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
oracledb
openpyxl